Se arreglaron problemas de puntuacion, y poder poner varios nombres y apellidos

+++ ChallengeApi/Entities/Pelicula.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Nombre { get; set; } = string.Empty;
         public int Duracion { get; set; }
-        public float Puntuacion { get; set; }
+        public decimal Puntuacion { get; set; }
         public decimal PresupuestoUsd { get; set; }
         // Relación con Productora
         public int ProductoraId { get; set; }
